requirements.txt
langchain
python-dotenv
langchain-openai
beautifulsoup4
langchain_community
langchain_ollama
faiss-cpu
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette